+++ core/src/integrationTest/kotlin/io/opentelemetry/android/AgentInitTest.kt
@@ -6,7 +6,6 @@
 package io.opentelemetry.android
 
 import android.app.Application
-import android.os.Build.VERSION_CODES.LOLLIPOP
 import android.os.Build.VERSION_CODES.UPSIDE_DOWN_CAKE
 import androidx.test.core.app.ApplicationProvider.getApplicationContext
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -18,7 +17,7 @@ import org.robolectric.annotation.Config
 
 @RunWith(AndroidJUnit4::class)
 internal class AgentInitTest {
-    @Config(sdk = [LOLLIPOP, UPSIDE_DOWN_CAKE])
+    @Config(sdk = [UPSIDE_DOWN_CAKE])
     @Test
     fun startOpenTelemetryRumInAndroid() {
         val application = getApplicationContext<Application>()