+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Drop support for Lollipop and API level 21.
+
 ## Version 0.14.0 (2025-08-21)
 
 ### 📣 Migration notes